Add option to disable sorting of layout
Lets the user specify that the layout should not be sorted by the layout engine, but instead let the filename control the order (ideal if making a spritesheet of an animation and the user wants the order of the sprites to be consistent with the animation sequence). Defaults to true to make i backwards compatible with previous behaviour.